pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "maskgit3d"
version = "0.1.0"
description = "A PyTorch + Lightning + MONAI deep learning framework with dependency injection"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["deep-learning", "pytorch", "lightning", "monai", "medical-imaging"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "lightning>=2.0.0",
    "monai>=1.0.0",
    "injector>=0.21.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "medmnist>=3.0.0",
    "einops>=0.7.0",
    "nibabel>=5.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "poetry"
]

[project.scripts]
maskgit3d-train = "maskgit3d.train:main"
maskgit3d-test = "maskgit3d.eval:main"

[project.urls]
Homepage = "https://github.com/yourusername/maskgit-3d"
Repository = "https://github.com/yourusername/maskgit-3d"
Issues = "https://github.com/yourusername/maskgit-3d/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/maskgit3d"]

[tool.hatch.build.targets.wheel.force-include]
"src/maskgit3d/conf" = "maskgit3d/conf"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501", "N802", "N812"]

[tool.ruff.lint.pep8-naming]
ignore-names = ["B", "D", "H", "W", "N", "C", "T", "L", "*Loss*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=maskgit3d --cov-report=term-missing --cov-report=xml"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require a GPU (deselect with '-m \"not gpu\"')",
    "integration: marks tests as integration tests",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
exclude = ["tests"]

[[tool.mypy.overrides]]
module = "medmnist"
ignore_missing_imports = true

[tool.pyright]
pythonVersion = "3.10"
venvPath = "/opt/homebrew/anaconda3/envs"
venv = "maskgit3d"
include = ["src"]
exclude = ["**/__pycache__"]
reportMissingImports = "error"
reportMissingTypeStubs = "warning"

[tool.coverage.run]
branch = true
source = ["src/maskgit3d"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "except ImportError:",
]
